requirements.txt
bs4==0.0.1
locust==1.1.1
pyyaml==5.3
